V2 Dirty Mnesia Support improvement

+++ test/test_helper.exs
@@ -13,11 +13,13 @@ use  Noizu.Database.Scaffolding.Test.Fixture.FooV2Table
 alias Noizu.Database.Scaffolding.Test.Fixture.FooTable
 alias Noizu.Database.Scaffolding.Test.Fixture.FooV2Table
 
-
-Amnesia.stop
-Amnesia.Schema.destroy
-Amnesia.Schema.create()
+if Mix.env == :test do
+  Amnesia.stop
+  Amnesia.Schema.destroy
+  Amnesia.Schema.create()
+end
 Amnesia.start
+
 if !Amnesia.Table.exists?(FooTable), do: FooTable.create(memory: [node()])
 if !Amnesia.Table.exists?(FooV2Table), do: FooV2Table.create(memory: [node()])
